requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
arxiv>=2.2.0
pymupdf>=1.24.0
google-genai>=1.22.0
python-dotenv>=1.1.1
httpx>=0.27.0
pydantic[email]>=2.11.7
